networking_6wind: use fp-vdev and fp-vdev-remote
fp-vdev utility is a part of virtual-accelerator product, it allows to
get proper vhostuser socket settings. At the same time,
neutron-fastpath-agent config contains only vhostuser section with the
same settings as well. So, let's use fp-vdev instead of supporting
parser code from networking_6wind/common/config.py.

In case of containerized setup like RedHat OSP, nova-compute, libvirtd,
neutron-fastpath-agent, virtual-accelerator can be installed and run in
different containers. Hence, we can't simply call fp-vdev to get
vhostuser settings or to create/delete vhostuser port, because fp-vdev
can be absent in these containers. As workaround, virtual-accelerator
package brings and starts fp-rpcd daemon. It serves as a proxy to bypass
fp-vdev requests from containers to fp-vdev binary in a host-hypervisor.
fp-rpcd.sock should be created in a folder shared between nova_compute
and neutron-fastpath-agent containers.

networking_6wind plugin package brings fp-vdev-remote client to fp-rpcd,
which connects to fp-rpcd.sock and sends requests to fp-vdev.

Moreover, vhostuser socket should be also created in a folder shared
between nova_libvirt container and host-hypervisor, because qemu process
is launching in nova_libvirtd container while creating a VM, but fp-rte
is running in the host.

To keep compatibility with any fastpath version and limit modifications,
a command to create or delete a virtual port is identical whatever the
binary is used (fp-vdev or fp-vdev-remote). Choice of the fp-vdev binary
is done by checking the presence of the fp-rpcd.sock file.

Change-Id: I529a313c082af9cb9ff9095a8fc0ba7bce10f3a2
Signed-off-by: Valentina Krasnobaeva <valentina.krasnobaeva@6wind.com>
Acked-by: Maxime Leroy <maxime.leroy@6wind.com>