common/constants: sort and rename some constants
* Sort constants in alphabetical order and rename some of them to have
more brief and clear names.

* Rename VIF_VHOSTUSER_FP_PLUG to VIF_DETAILS_VHOSTUSER_FP_PLUG as it
called now in nova.network.model.py. We can't simply import this
constant from nova.network module instead of duplicate it, because in
case of containerized setups networking_6wind can be installed in
containers, that can't access to nova libraries.

* Tests: add portbindings.VIF_TYPE_TAP to supported plug methods list,
this type is always presented in this list with all supported version of
VA.

Change-Id: I884dcdc14061c585c27d6c77c09306d53a838816
Signed-off-by: Valentina Krasnobaeva <valentina.krasnobaeva@6wind.com>
Acked-by: Maxime Leroy <maxime.leroy@6wind.com>